backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
psycopg2-binary
